Moves all item type classes to one single file
